Convert ride groups to C++
