--- src/openrct2/ride/ride_group.h
@@ -1,47 +1,0 @@
-#pragma region Copyright (c) 2014-2017 OpenRCT2 Developers
-/*****************************************************************************
- * OpenRCT2, an open source clone of Roller Coaster Tycoon 2.
- *
- * OpenRCT2 is the work of many authors, a full list can be found in contributors.md
- * For more information, visit https://github.com/OpenRCT2/OpenRCT2
- *
- * OpenRCT2 is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * A full copy of the GNU General Public License can be found in licence.txt
- *****************************************************************************/
-#pragma endregion
-
-#ifndef _RIDE_GROUP_H
-#define _RIDE_GROUP_H
-
-#include "../common.h"
-#include "ride_data.h"
-
-typedef struct ride_group {
-    uint8 track_type;
-    uint16 maximum_height;
-    uint64 available_track_pieces;
-    rct_ride_name naming;
-} ride_group;
-
-ride_group * get_ride_group(uint8 trackType, rct_ride_entry * rideEntry);
-bool track_type_has_ride_groups(uint8 trackType);
-void generate_ride_groups();
-ride_group * ride_group_generate_fallback(uint8 trackType);
-ride_group * ride_group_find(uint8 rideType, uint8 index);
-bool ride_groups_are_equal(ride_group * a, ride_group * b);
-bool ride_group_is_invented(ride_group * rideGroup);
-
-extern ride_group ride_group_corkscrew_rc;
-extern ride_group ride_group_hypercoaster;
-extern ride_group ride_group_car_ride;
-extern ride_group ride_group_monster_trucks;
-extern ride_group ride_group_steel_twister_rc;
-extern ride_group ride_group_hyper_twister;
-extern ride_group ride_group_junior_rc;
-extern ride_group ride_group_midi_coaster;
-
-#endif
