+++ src/openrct2/ride/track.c
@@ -34,10 +34,10 @@
 #include "ride.h"
 #include "ride_data.h"
 #include "ride_ratings.h"
+#include "RideGroup.h"
 #include "station.h"
 #include "track.h"
 #include "track_data.h"
-#include "ride_group.h"
 
 uint8 gTrackGroundFlags;
